% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xv.R
\name{folded.Xv}
\alias{folded.Xv}
\title{Folded Matrix-Vector Multiplication}
\arguments{
\item{X}{Matrix.}

\item{v}{numeric vector.}

\item{foldid}{integer vector. Specifying the foldid of each rows.}

\item{target}{integer value. The target fold id}

\item{is_exclude}{logical value. Whether to exclude those rows belong to the target fold or not.}
}
